requirements.txt
colour==0.1.5
imageio==2.33.1
matplotlib==3.8.2
numpy==1.23.3
pandas==1.4.4
Pillow==10.2.0
sympy==1.12